Execution_engine: make `Do_make` jobs work
As explained in the comments, this makes them work but there is much
room for optimization, since the actions related to a `make` job are
started by batches, and every time we wait for all the actions to finish
before starting the next batch. At least it seems to work.